feat: Create files in a GitLab project
This replicates what we do in the GitHub equivalent module, where we
provide a variable to create files in a repository/project. These
files can either be "managed", where Terraform puts them in the state
and will overwrite any changes the next time it runs, or "unmanaged",
where Terraform does a one-time create action.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>